Fix VFS-359 - Don't delete a RamFileObject if it is open
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/trunk@1175302 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/commons/vfs2/provider/ram/RamFileObject.java
@@ -135,6 +135,10 @@ public class RamFileObject extends AbstractFileObject implements FileObject
     @Override
     protected void doDelete() throws Exception
     {
+
+        if (this.isContentOpen()) {
+            throw new FileSystemException(this.getName() + " cannot be deleted while the file is openg");
+        }
         fs.delete(this);
     }
